requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
numpy==1.16.3
opencv-python==4.1.0.25
paho-mqtt==1.4.0
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
